Revert "Students service used successfully"
This reverts commit 223741804f0a1643fdae777704bdaf93c110d470.

This was a big mistake, I created a second branch to implement
routing so that when I click on a course it goes to a different page instead of displaying in the same page. It was a big change as the communication
between parent and son component dissapears and those components
become 'siblings' instead. The mistake was that when I was trying to implement
the routes I created the app-routing.module.ts file and the original
branch had the app.module.ts with the routes within. Reason why I want
to add the new routing file indeed but first in the main branch and then
the new branch can be merged correctly.